protocol-kit shouldn't be a dep of SafeOperation
