+++ packages/relay-kit/src/packs/safe-4337/SafeOperationFactory.ts
@@ -1,4 +1,3 @@
-import Safe from '@safe-global/protocol-kit'
 import {
   UserOperation,
   UserOperationV06,
@@ -13,13 +12,12 @@ import { isEntryPointV7 } from '@safe-global/relay-kit/packs/safe-4337/utils'
 class SafeOperationFactory {
   static createSafeOperation(
     userOperation: UserOperation,
-    protocolKit: Safe,
     options: SafeOperationOptions
   ): SafeOperationBase {
     if (isEntryPointV7(options.entryPoint)) {
-      return new SafeOperationV07(userOperation as UserOperationV07, protocolKit, options)
+      return new SafeOperationV07(userOperation as UserOperationV07, options)
     } else {
-      return new SafeOperationV06(userOperation as UserOperationV06, protocolKit, options)
+      return new SafeOperationV06(userOperation as UserOperationV06, options)
     }
   }
 }
